requirements.txt
setuptools>=49.2.1
discord~=1.0.1
disnake~=2.5.2
requests>=2.25.1
googletrans~=4.0.0rc1
bs4~=0.0.1
beautifulsoup4~=4.11.1
google-generativeai>=0.3.2